docker/dashboard/requirements.txt
flask
flask-socketio
flask-cors
eventlet
paho-mqtt>=1.6.1
requests
python-socketio[asyncio,eventlet]>=5.8.0
